+++ react/__tests__/ProductSummaryImage.test.js
@@ -100,4 +100,49 @@ describe('<ProductImage /> component', () => {
 
     expect(svgElements).toHaveLength(1)
   })
+
+  it('should render custom placeholder img', () => {
+    const rawProduct = {
+      productId: '123456789',
+      linkText: 'linkText',
+      productName: 'productName',
+      items: [
+        {
+          itemId: '1',
+          name: 'item 1',
+          sellers: [
+            {
+              sellerId: '1',
+              commertialOffer: {
+                AvailableQuantity: 0,
+                Price: 10,
+                ListPrice: 7,
+              },
+            },
+          ],
+          images: [],
+        },
+      ],
+      productClusters: [
+        {
+          name: 'name',
+        },
+      ],
+      quantity: 1,
+    }
+    const placeholder = 'placeholder-image-url'
+
+    const product = ProductSummary.mapCatalogProductToProductSummary(rawProduct)
+
+    const { container } = render(
+      <ProductSummary product={product}>
+        <ProductImage placeholder={placeholder} />
+      </ProductSummary>
+    )
+
+    const imgElements = container.getElementsByTagName('img')
+
+    expect(imgElements).toHaveLength(1)
+    expect(imgElements[0].src.includes(placeholder)).toBeTruthy()
+  })
 })