+++ react/components/ProductSummaryImage/ProductImage.js
@@ -138,6 +138,7 @@ const ProductImageContent = ({
   mainImageLabel,
   hoverImageLabel,
   showCollections,
+  placeholder,
   width: widthProp,
   height: heightProp,
   aspectRatio,
@@ -172,7 +173,9 @@ const ProductImageContent = ({
 
   let skuImageUrl = pathOr('', ['image', 'imageUrl'], sku)
 
-  if (!skuImageUrl || hasError) {
+  const shouldDisplayPlaceholder = !skuImageUrl || hasError
+
+  if (!placeholder && shouldDisplayPlaceholder) {
     return (
       <div className={containerClassname}>
         <ImagePlaceholder cssHandle={handles.productImage} />
@@ -220,7 +223,7 @@ const ProductImageContent = ({
   const img = (
     <div className={containerClassname}>
       <Image
-        src={skuImageUrl}
+        src={shouldDisplayPlaceholder ? placeholder : skuImageUrl}
         width={width}
         height={height}
         aspectRatio={aspectRatio}
@@ -254,6 +257,7 @@ const ProductImage = ({
   mainImageLabel,
   hoverImageLabel,
   showCollections,
+  placeholder,
   width: widthProp,
   height: heightProp,
   aspectRatio: aspectRatioProp,
@@ -295,6 +299,7 @@ const ProductImage = ({
         mainImageLabel={mainImageLabel}
         hoverImageLabel={hoverImageLabel}
         showCollections={showCollections}
+        placeholder={placeholder}
       />
     </div>
   )
@@ -311,6 +316,7 @@ ProductImage.propTypes = {
   displayMode: PropTypes.oneOf(['normal', 'inline']),
   hoverImageLabel: PropTypes.string,
   mainImageLabel: PropTypes.string,
+  placeholder: PropTypes.string,
   width: PropTypes.oneOfType([PropTypes.number, PropTypes.object]),
   height: PropTypes.oneOfType([PropTypes.number, PropTypes.object]),
 }