+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ProductSummaryImage` can now receive a custom placeholder image by using the `placeholder` prop.
 
 ## [2.63.1] - 2020-11-13
 ### Changed